moved logged_user to base conftest
